% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqplot.R
\name{cq.plot}
\alias{cq.plot}
\title{Plot Relative Expression Values}
\usage{
cq.plot(tablename, groupColumn = "Sample", plotWidth = 10,
  plotHeight = 10)
}
\arguments{
\item{tablename:}{a cqimport object}

\item{groupColumn:}{defaults to "Sample". Defines the grouping column which will define coloring.}

\item{plotWidth:}{width of output pdf file}

\item{plotHeight:}{height of output pfd file}
}
\description{
This function plots the untransformed Cq values for each target.
}
\examples{

df <- cqimport(filename)
p1 <- cq.plot(df)

}
\keyword{expression,}
\keyword{plot}
\keyword{qPCR,}
\keyword{relative}
