% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cqimport.R
\name{cqimport}
\alias{cqimport}
\title{Import BioRad qPCR Data}
\usage{
cqimport(indir, tablename, dropunnecessary = TRUE)
}
\arguments{
\item{tablename:}{This is the filename of the .txt file containing Cq values as copied from the file manager.}

\item{indir:}{The path of the root directory containing the script.}

\item{dropunnecessary:}{Defaults to TURE. If TRUE, returns only the columns "Well", "Target", "Sample", "Cq".}
}
\description{
This function imports a .txt file from CFX manager and returns a dataframe.
}
\details{
First simply use the file name (without .txt) as input, assuming subfolder /input as source.
The Cq is set to 40 in those wells were the threshold was not reached.
}
\examples{
#The .txt files should be in: "parent_directory/input/example_file_name.txt"

setwd("parent_directory")
filename <- "example_file_name"
df <- cqimport(filename)
}
\keyword{import}
\keyword{qPCR,}
