% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meltplot.R
\name{mc.plot}
\alias{mc.plot}
\title{Plot Relative Expression Values}
\usage{
mc.plot(tablename, groupColumn = "Sample", plotWidth = 10,
  plotHeight = 10)
}
\arguments{
\item{tablename:}{an mcimport object}

\item{groupColumn:}{defaults to "Sample". Defines the grouping column which will define coloring.}

\item{plotWidth:}{width of output pdf file}

\item{plotHeight:}{height of output pfd file}
}
\description{
This function plots the meltcurves for each target.
}
\examples{

setwd("parent_directory")
filename <- "example_file_name"
meltderivative <- "example_file_name"

df <- Rsome::cqimport(filename)
p1 <- Rsome::cq.plot(df)
mc <- Rsome::mcimport(cqimport = df, meltderivative = meltderivative)
p2 <- Rsome::mc.plot(mc)

}
\keyword{expression,}
\keyword{plot}
\keyword{qPCR,}
\keyword{relative}
