% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcimport.R
\name{mcimport}
\alias{mcimport}
\title{Import BioRad qPCR Data}
\usage{
mcimport(cqimport = df, meltderivative = meltderivative)
}
\arguments{
\item{cqimport:}{A cqimport object.}

\item{meltderivative:}{This is the file containing the meltcurve derivative data.}
}
\description{
This function imports a meltcurve derivative .txt file from CFX manager and returns a dataframe.
}
\details{
First simply use the file name (without .txt) as input, assuming subfolder /input as source.
}
\examples{
#The .txt files should be in: "parent_directory/input/example_file_name.txt"

setwd("parent_directory")
filename <- "example_file_name"
meltderivative <- "example_file_name"
df <- cqimport(filename)
mc <- mcimport(cqimport = df, meltderivative = meltderivative)
}
\keyword{import}
\keyword{qPCR,}
