% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replot.R
\name{re.plot}
\alias{re.plot}
\title{Plot Relative Expression Values}
\usage{
re.plot(tablename, groupColumn = "Sample")
}
\arguments{
\item{tablename:}{a relex object}

\item{groupColumn:}{defaults to "Sample". Defines the grouping column which will define coloring.}
}
\description{
This function plots the relative expression values for each target normalized to the household gene.
}
\examples{

df <- cqimport(filename)
re <- relex(df, household = "Gapdh")
p1 <- re.plot(re)

}
\keyword{expression,}
\keyword{plot}
\keyword{qPCR,}
\keyword{relative}
