% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relex.R
\name{relex}
\alias{relex}
\title{Calculate The 2^dCt}
\usage{
relex(tablename, household = "Gapdh")
}
\arguments{
\item{tablename:}{a cqimport object}

\item{household:}{defaults to "Gapdh". Accepts a string to define the household gene.}
}
\description{
This function calculates the relative expression from a cqimport object.
}
\examples{

df <- cqimport(filename)
rel.expression <- relex(df, household = "Gapdh")

}
\keyword{expression}
\keyword{qPCR,}
\keyword{relative}
